Rename types/file_object.py to types/blob_storage.py
