+++ packages/syft/src/syft/service/blob_storage/stash.py
@@ -3,7 +3,7 @@ from ...serde.serializable import serializable
 from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionSettings
-from ...types.file_object import FileObject
+from ...types.blob_storage import FileObject
 
 
 @serializable()
