+++ packages/syft/src/syft/store/blob_storage/__init__.py
@@ -54,9 +54,9 @@ from ...serde.serializable import serializable
 from ...service.response import SyftError
 from ...service.response import SyftSuccess
 from ...types.base import SyftBaseModel
-from ...types.file_object import CreateFileObject
-from ...types.file_object import FileObject
-from ...types.file_object import SecureFilePathLocation
+from ...types.blob_storage import CreateFileObject
+from ...types.blob_storage import FileObject
+from ...types.blob_storage import SecureFilePathLocation
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
