+++ packages/syft/src/syft/types/blob_storage.py
